♻️  refactor: relocate NotFoundPage to error folder
